fix search except for armourClass

+++ src/main/java/com/dnd/tools/encounterhelper/monster/MonsterSearch.java
@@ -15,7 +15,7 @@ public class MonsterSearch {
   @Data
   @AllArgsConstructor
   static class Range {
-    private int lowerBound;
-    private int upperBound;
+    private double lowerBound;
+    private double upperBound;
   }
 }
